fix styles not being applied to new tabs opened
When opening a new tab (command-click or right clicking on a link), the styles
would not load on the new tab. This appears to be due to the tab not counting as
a "primary frame" (perhaps this does not completely refer to iframes?). Also
this might fix issue with iframes too, who knows. I don't have a great test for
iframes right now.