fix various issues
It looks like when I hoped on the develop branch, I encountered some
result of refactoring that was not complete yet. I may dress this up
later and send it to the author as a PR.

Some things of note:
I might have taken short cuts to establishing a version 3 payload.
- The @import comments seem to be completely ignored.
- Some error handling was introduced for more verbosity during an error.
- Iframes still don't work, because Security.